*** empty log message ***

+++ coders/preview.c
@@ -401,7 +401,7 @@ static unsigned int WritePreviewImage(const ImageInfo *image_info,Image *image)
       case ThresholdPreview:
       {
         FormatString(factor,"%lu",(unsigned long)
-          ((percentage*(MaxRGB+1))/100));
+          ((percentage*((double) MaxRGB+1.0))/100));
         FormatString(label,"threshold %.1024s",factor);
         commands[argc++]=(char *) "-threshold";
         commands[argc++]=factor;
